(#115) refactor : 칵테일, 바 Prop 수정
