+++ src/libs/utils/Bardetaildummy.ts
@@ -19,34 +19,28 @@ export const BarInfo: BarProps = {
   coverCharge: "5000",
   barsCocktail: [
     {
-      title: "임시",
-      img: cocktailImg1,
-      level: "수준",
-      description: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
+      cocktailName: "임시",
+      cocktailPicture: cocktailImg1,
+      recoUser: 2,
+      cocktailDetail: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
     },
     {
-      title: "임시",
-      img: cocktailImg2,
-      level: "수준",
-      description: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
+      cocktailName: "임시",
+      cocktailPicture: cocktailImg1,
+      recoUser: 0,
+      cocktailDetail: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
     },
     {
-      title: "임시",
-      img: cocktailImg3,
-      level: "수준",
-      description: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
+      cocktailName: "임시",
+      cocktailPicture: cocktailImg1,
+      recoUser: 1,
+      cocktailDetail: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
     },
     {
-      title: "임시",
-      img: cocktailImg3,
-      level: "수준",
-      description: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
-    },
-    {
-      title: "임시",
-      img: cocktailImg2,
-      level: "수준",
-      description: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
+      cocktailName: "임시",
+      cocktailPicture: cocktailImg1,
+      recoUser: 2,
+      cocktailDetail: `대표적인 레이디 킬러 칵테일 중 하나로,\n 달콤한 맛에 비해 도수가 무려 20도를 넘어선다.\nIBA 공식 칵테일에 등록되어 있는 레시피 중 하나`,
     },
   ],
 };
